+++ src/parser/instruction/itype.rs
@@ -6,7 +6,7 @@ pub enum IType {
     Lb,
 }
 impl IType {
-    pub fn execute(&self, cpu: &mut Processor, rd: i32, rs: i32, imm: i32) {
+    pub fn execute(&self, cpu: &mut Processor, rd: usize, rs: usize, imm: i32) {
         match self {
             IType::Addi => cpu.registers[rd as usize] = cpu.registers[rs as usize] + imm,
             IType::Lb => {
